+++ demo/src/DemoTagList.tsx
@@ -1,7 +1,6 @@
-"use strict";
-
 const debug = require("debug")("DemoTagList");
 
+import autobind from "autobind-decorator";
 import * as React from "react";
 import {Break, TagList} from "../../dist/bundle";
 import {StyledContainer} from "./helpers";
@@ -12,12 +11,24 @@ export default class DemoTagList extends React.Component<any, undefined> {
 		debug("creating");
 	}
 
+	@autobind
+	private handleNew(tag: string, tags: string[]) {
+		debug(`handleNew -> tag: %o, tags: %O`, tag, tags);
+	}
+
+	@autobind
+	private handleDelete(tag: string, tags: string[]) {
+		debug(`handleDelete -> tag: %o, tags: %O`, tag, tags);
+	}
+
 	public render() {
 		return (
 			<StyledContainer id='tagListExample' title='TagList'>
 				<h3>Static</h3>
 				<TagList
 					disabled={this.props["disabled"]}
+					onDelete={this.handleDelete}
+					onNew={this.handleNew}
 					sizing={this.props["sizing"]}
 					tags={["one", "two", "three"]}
 				/>
@@ -27,6 +38,8 @@ export default class DemoTagList extends React.Component<any, undefined> {
 				<TagList
 					disabled={this.props["disabled"]}
 					nosort
+					onDelete={this.handleDelete}
+					onNew={this.handleNew}
 					sizing={this.props["sizing"]}
 					tags={["aaa", "ccc", "bbb"]}
 					useinput
@@ -36,6 +49,8 @@ export default class DemoTagList extends React.Component<any, undefined> {
 				<h3>Changeable (sorted)</h3>
 				<TagList
 					disabled={this.props["disabled"]}
+					onDelete={this.handleDelete}
+					onNew={this.handleNew}
 					sizing={this.props["sizing"]}
 					tags={["aaa", "ccc", "bbb"]}
 					useinput